feat: 메인 화면에서 카트 상품 수량 감소, 삭제 구현

+++ app/src/main/java/woowacourse/shopping/data/carts/CartDao.kt
@@ -35,7 +35,7 @@ interface CartDao {
     fun update(cartEntity: CartEntity)
 
     @Query("SELECT * FROM cart WHERE id = :id")
-    fun findByGoodsId(id: Long): CartEntity?
+    fun findByGoodsId(id: Int): CartEntity?
 
     @Transaction
     fun addOrIncreaseQuantity(cartEntity: CartEntity) {
