+++ app/src/main/java/woowacourse/shopping/data/util/mapper/toDomain.kt
@@ -4,4 +4,4 @@ import woowacourse.shopping.data.goods.dto.Content
 import woowacourse.shopping.domain.model.Goods
 
 fun Content.toDomain(): Goods =
-    Goods(this.name,this.price,this.imageUrl,this.id.toLong(),this.category)
+    Goods(this.name,this.price,this.imageUrl,this.id,this.category)
